move setup secrets
